requirements.txt
click==6
temperusb
requests
flask
psycopg2
python-dateutil